storage_writer/pyproject.toml
[project]
name = "synthetic-ledger-storage-writer"
version = "0.1.0"
description = "Fan-out writer for ClickHouse, Druid and Pinot"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "httpx>=0.27.0",
  "opentelemetry-proto>=1.27.0",
  "orjson>=3.10.0",
  "protobuf>=5.28.0",
  "pydantic>=2.8.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
